Create and register application blueprints
