+++ back_end/views/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+app_views = Blueprint('app_views', __name__)
+
+from views.auth_routes import *
